fix(auth components): auth components fix
