+++ src/components/shared/Input/Input.tsx
@@ -1,8 +1,5 @@
-import React, { useState } from "react";
+import React from "react";
 import styled from "styled-components";
-
-import { passwordValidator } from "../../../utils/validator/validator";
-import { emailValidator } from "../../../utils/validator/validator";
 
 export const StyledInput = styled.input`
   display: block;
@@ -18,32 +15,14 @@ export const StyledInput = styled.input`
 interface Props extends React.HTMLProps<HTMLInputElement> {
   type: string;
   placeholder: string;
-  handleChange: (inputType: any, value: any) => void;
+  handleChange: (event: any) => void;
 }
 
 export const Input: React.FC<Props> = ({ type, placeholder, handleChange }) => {
-  const isValid = (inputValue: string, inputType: string) => {
-    if (inputType === "email") {
-      return emailValidator(inputValue);
-    } else if (inputType === "password") {
-      return passwordValidator(inputValue);
-    }
-  };
-
-  const handleOnChange = (event: any) => {
-    const value = event.target.value;
-    const inputType = event.target.type;
-    if (isValid(value, inputType)) {
-      handleChange(inputType, value);
-    } else {
-      handleChange("", "");
-    }
-  };
-
   return (
     <StyledInput
       type={type}
-      onChange={handleOnChange}
+      onChange={handleChange}
       placeholder={placeholder}
     />
   );
